requirements.txt
# ── Web Framework ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ── Task Queue ─────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.4

# ── Storage ─────────────────────────────────────────────────────────
minio==7.2.7

# ── Auth ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Config / Validation ─────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0

# ── Logging ─────────────────────────────────────────────────────────
structlog==24.2.0

# ── Computer Vision ─────────────────────────────────────────────────
opencv-python-headless==4.9.0.80
Pillow==10.3.0
scikit-image==0.23.2
imutils==0.5.4

# ── Deep Learning ───────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0

# ── ML Utilities ────────────────────────────────────────────────────
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
albumentations==1.4.8
tqdm==4.66.4
matplotlib==3.9.0
seaborn==0.13.2

# ── Testing ─────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0
flower==2.0.1
ultralytics==8.2.0